+++ R/pumpCase.R
@@ -27,11 +27,11 @@ pumpCase.euclidean <- function(x, case = "address") {
 }
 
 #' @export
-pumpCase.voronoi <- function(x, case = "address") {
+pumpCase.voronoi <- function(x, case = "orthogonal") {
   output <- x$statistic.data
-  if (x$case.location == "address") {
+  if (x$case.location == "orthogonal") {
     lapply(output, function(x) cholera::ortho.proj$case[x == 1])
-  } else if (x$case.location == "anchor") {
+  } else if (x$case.location == "nominal") {
     lapply(output, function(x) cholera::fatalities.address$anchor[x == 1])
   }
 }