+++ NEWS.md
@@ -1,3 +1,11 @@
+### cholera 0.8.0.9366
+
+- use neighborhoodVoronoi() as wrapper for voronoiNominal() and 
+  voronoiLatlong().
+- set "nominal" and "orthogonal" as arguments for case.location and 
+  pump.location.
+
+
 ### cholera 0.8.0.9365
 
 - use neighborhoodEuclidean() as wrapper for euclideanNominal() and